+++ examples/threejs/vite.config.js
@@ -1,0 +1,5 @@
+// vite.config.js
+export default {
+  base: './',
+};
+