+++ src/squint/internal/cli.cljs
@@ -10,7 +10,8 @@
    [squint.repl.node :as repl]
    [squint.repl.nrepl-server :as nrepl]
    [squint.internal.node.utils :as utils]
-   [clojure.string :as str]))
+   [clojure.string :as str])
+  (:require-macros [squint.resource :refer [version]]))
 
 (defn file-in-output-dir [file paths output-dir]
   (path/resolve output-dir
@@ -74,7 +75,7 @@
               files))))
 
 (defn print-help []
-  (println "Squint v0.0.0
+  (println (str "Squint v" (version) "
 
 Usage: squint <subcommand> <opts>
 
@@ -87,7 +88,7 @@ compile   <file.cljs> ... Compile file(s)
 repl                      Start repl
 help                      Print this help
 
-Use squint <subcommand> --help to show more info."))
+Use squint <subcommand> --help to show more info.")))
 
 (defn fallback [{:keys [rest-cmds opts]}]
   (if-let [e (:e opts)]